Split MessageList into two components
